pyproject.toml
[project]
name = "genai-misc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arxiv>=2.2.0",
    "dspy>=2.6.11",
    "duckdb>=1.3.1",
    "ipykernel>=6.29.5",
    "kagglehub>=0.3.12",
    "langchain>=0.3.25",
    "langchain-openai>=0.3.18",
    "litellm>=1.63.6",
    "matplotlib>=3.10.1",
    "mcp[cli]>=1.13.1",
    "mlx>=0.23.2",
    "mlx-lm>=0.21.5",
    "numpy>=2.2.3",
    "openai>=1.66.0",
    "pandas>=2.2.3",
    "polars>=1.24.0",
    "pyyaml>=6.0.2",
    "sentence-transformers>=4.1.0",
]
